dom/src/threads/test/test_threadErrors.html

<!DOCTYPE HTML>
<html>
<!--
Tests of DOM Worker Threads (Bug 437152)
-->
<head>
  <title>Test for DOM Worker Threads (Bug 437152)</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=437152">DOM Worker Threads Bug 437152</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

  netscape.security.PrivilegeManager.enablePrivilege('UniversalXPConnect');

  var prefs = Components.classes["@mozilla.org/preferences-service;1"].
              getService(Components.interfaces.nsIPrefBranch);
  //prefs.setIntPref("javascript.options.gczeal", 2);

  var badWorkerScripts = [
    // Syntax error
    "function messageListener(message, source) { " +
    "  for (var i = 0; i < 10) { } " +
    "} " +
    "",

    // Bad function error
    "function messageListener(message, source) { " +
    "  foopy(); " +
    "} " +
    "",

    // Unhandled exception in body
    "function messageListener(message, source) { " +
    "} " +
    "throw new Error('Bah!'); " +
    "",

    // Throwing message listener
    "function messageListener(message, source) { " +
    "  throw 'Bad message: ' + message; " +
    "} " +
    ""
  ];

  var expectedErrorCount = badWorkerScripts.length;

  var pool = navigator.newWorkerPool();

  pool.messageListener = function(message, source) {
    netscape.security.PrivilegeManager.enablePrivilege('UniversalXPConnect');

    prefs.setIntPref("javascript.options.gczeal", 0);
    ok(false, "Unexpected message");
    SimpleTest.finish();
  };

  var actualErrorCount = 0;
  var failedWorkers = [];

  pool.errorListener = function(error, source) {
    netscape.security.PrivilegeManager.enablePrivilege('UniversalXPConnect');

    if (failedWorkers.indexOf(source) != -1) {
      dump("Already seen worker: " + source + "\n");
      ok(false, "Seen an extra error from this worker");
      prefs.setIntPref("javascript.options.gczeal", 0);
      SimpleTest.finish();
      return;
    }

    failedWorkers.push(source);
    actualErrorCount++;

    if (actualErrorCount == expectedErrorCount) {
      prefs.setIntPref("javascript.options.gczeal", 0);
      SimpleTest.finish();
    }
  };

  for (var i = 0; i < expectedErrorCount; i++) {
    var worker = pool.createWorker(badWorkerScripts[i]);
    worker.postMessage("Hi");
  }

  SimpleTest.waitForExplicitFinish();
</script>
</pre>
</body>
</html>
